+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Removed `excluded_fields` deprecated configuration attribute for an entity. Use `exclude` attribute for a field instead.
 * Removed `result_name` deprecated configuration attribute for a field. Use `property_path` attribute instead.
 * Removed `orderBy` deprecated configuration attribute. Use `order_by` attribute instead.
+* Removed deprecated signature `function (array &$item) : void` of post serialization handler that can be specified in `post_serialize` configuration attribute. Use `function (array $item, array $context) : array` instead.
 
 ### Changed
 #### AddressBundle